refactor cleaner
